[bugfix] Converted long double to double in `frexpl`.
This is to avoid calls to libgcc soft-fp with which we are binary incompatible.

+++ include/math.h
@@ -230,7 +230,7 @@ long double fmodl(long double, long double);
 
 double      frexp(double, int *);
 float       frexpf(float, int *);
-long double frexpl(long double, int *);
+double frexpl(double, int *);
 
 double      hypot(double, double);
 float       hypotf(float, float);
